Merged PR 850868: Update controls library color dependency
This PR updates the control library to obtain colors from the file specified by the webpack sass-loader. This change allows the controls library to be able to see any overrides made in the baseline app. It is necessary to allow colors used in the controls library to be overwritten by color overrides specified in an external app, like the baseline app.

For the control library docs, it naturally falls back to directly use the colors from the css library.