Fix DatePicker + DateTimeField to return null and set value to empty when the value is removed (#86)
When we remove the value entirely from the date field, the value being returned in 'invalid' which in a sense breaks the validation as it is not a date expected. Adding a fix to return the value as null in order to let validation succeed as a empty string in the date field is essentially right.